requirements.txt
appdirs==1.4.4
attrs==19.3.0
bcrypt==3.2.0
black==19.10b0
blinker==1.4
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.0
distro==1.5.0
docker==4.3.0
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
Flask==1.1.2
Flask-Mail==0.9.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
magicmock==0.3
MarkupSafe==1.1.1
mock==4.0.2
more-itertools==8.4.0
packaging==20.4
paramiko==2.7.1
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
pytest-mock==3.2.0
python-dotenv==0.14.0
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
texttable==1.6.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.1.0
